feat(tracing): Add `tracePropagationTargets` option to browser routing instrumentation (#6080)
As part of the work on https://github.com/getsentry/sentry-javascript/issues/5285, this adds a new option, `tracePropagationTargets`, to our browser-side tracing instrumentation, to live alongside (and eventually take the place of) `tracingOrigins`. Its behavior matches that of `tracingOrigins`, but it has a much clearer name.

Switching from internal use of `tracingOrigins` to `tracePropagationTargets` to come in future PRs.

Note: This is what remained of https://github.com/getsentry/sentry-javascript/pull/6041 after the fix in https://github.com/getsentry/sentry-javascript/pull/6079 was split off. h/t @timfish for the initial work in that PR.