fix(nextjs): Allow `onUncaughtException` integration to remain excluded (#6148)
In the nextjs SDK, the `addOrUpdateIntegration` function exists to force the inclusion of certain integrations with certain options set. If such an integration is included in the underlying SDK's default integrations, however, (in other words, if it's included in the defaults set by `@sentry/browser` or `@sentry/node`), it's possible for the user to have actively chosen to exclude it, which we would then be overriding. This PR adds to the `addOrUpdateIntegration` logic to provide the ability to respect that choice.

The only way for a user to choose to filter out a default integration is by providing a function as their `integrations` option in `Sentry.init()`. Therefore, when handling the function case,  we can check if a given integration is included in the return value, and if it's not, not add the default instance we otherwise would. This is controlled by a flag on that default instance named `allowExclusionByUser`. If it's set to `true`, we'll perform the check and respect the user's choice. If it's set to `false` or not set at all, we'll continue to behave as we have, forcing the inclusion of the given integration.

The inspiration for this change is our recent inclusion of the `onUncaughtException` integration in the nextjs defaults. This PR therefore also applies the above change to that default instance.

Finally, the test suite for `addOrUpdateIntegration` has been entirely reworked, to ensure that it covers all possible cases.