fix(otel): Remove sampling decision in propagator (#6152)
Previously the Sentry Propagator was setting the sample rate on the opentelemetry span context based on the incoming traceparent from `sentry-trace`. This is problematic because it a span is unsampled, it does not enter a span processor. This means that we could create transactions erroneously (for example, unnecessarily promote a child span to a transaction).