chore(ember): Make tests less brittle (#6164)
We already had some logic to make test less brittle, but the runloop is a bit unstable, and thus hard to test. So it is easier to just don't test the runloop execution in detail, especially since these are the least interesting spans, and focus on testing the important parts (that are also more stable).