feat(util): Add `stringMatchesSomePattern` helper function. (#6205)
This adds a new utility function, `stringMatchesSomePattern`, which tests a string against an entire array of strings and/or regexes at once (which is something we do in a number of spots in the SDK). It also adds an optional parameter to `isMatchingPattern` to allow the string matching to be exact rather than a substring match.