chore(serverless): Fix incorrect path in Readme (#6212)
This patch fixes an incorrect path for requiring the auto-init script in the `@sentry/serverless` Readme for users who want to set up the layer via the ARN and environment variables. Furthermore, this patch slightly enhances the step to find the correct ARN by linking to our (still kinda broken) ARN dropdown in the docs.