fix(remix): Resolve Remix Request API compatibility issues. (#6215)
Ref: #6139 and https://github.com/getsentry/sentry-javascript/issues/6139#issuecomment-1315799722

This PR is not a complete solution for all issues mentioned in #6139, but it aims to solve the parsing issue, which should fix auto-instrumented usage.

Remix uses its own implementation of Fetch API [1] on both back-end and front-end, which has different structures storing `headers` and `url`. That has caused `RequestData` integration to not work properly on Remix projects.

I first attempted adding support to the core parser, and `PolymorphicRequest`. But it seemed very complex (if possible), because we have to type Proxy objects and Symbols, which our current TypeScript version doesn't fully support.

So, I vendored / modified a couple of unexported utilities from `@remix-run/web-fetch` [2], to convert request objects to a compatible structure that we can consume in `RequestData` integration.

[1] https://github.com/remix-run/web-std-io/tree/main/packages/fetch
[2] https://www.npmjs.com/package/@remix-run/web-fetch