fix(core): Include `_sdkProcessingMetadata` when cloning scope (#6218)
When I added the `_sdkProcessingMetadata` field to `Scope`, I missed adding it to the `clone` method. This fixes that.

Addresses the `withScope` problem mentioned in https://github.com/getsentry/sentry-javascript/issues/6139#issuecomment-1310839406. (TL;DR - the stored request upon which the `RequestData` integration relies was getting lost when `withScope` was called, which in turn meant that no request data was coming through.)