chore(replay): Add size limit entries for Replay (#6288)
Add size limit entries for Replay. One entry for the (currently still exisiting) NPM bundle and two more for the package built with WebPack.

When comparing the size-limit results here with the results from the replay repo, the size goes up from 35 to 83KB but I'm pretty sure our result is more representative (the file already has 135 KB when inspecting it locally). My suspicion is that the size-limit action from the replay repo measures size differently, as it uses other GH actions and newer size-limit versions.