chore(replay): Cleanup `package.json` scripts (#6348)
Clean up the replay integration's `package.json` by getting rid of a few ToDos, enabling the circular dependency check and using the eslint cache of the monorepo. 

Since the future of the demo app in the replay package is still unclear, I opted to repurpose and rename the `bootstrap` script so that it now only bootstraps the demo app. For the package itself, bootstrapping happens via a monorepo-level `yarn` command or `lerna bootstrap` command.