feat(nextjs): Parameterize prefix loader values (#6377)
This changes the `prefixLoader` we use in the nextjs SDK to accept template name and replacement values as parameters, and pulls injection of that loader into a helper function. It also renames the current template from `prefixLoaderTemplate` to `serverRewriteFramesPrefixLoaderTemplate`, to better indicate its purpose and to distinguish it from future templates also using the prefix loader.