build(replay): Generate Replay CDN bundles (#6391)
Add a rollup config to generate an addon CDN bundle for the Replay integration. Usage of this bundle is identical to our other integration CDN bundles in the sense that the bundle has to be added in addition to the regular browser (or browser+tracing) bundle. 

As of this PR, three bundles will be generated into `build/bundles`:

* `replay.js` (the unminified bundle with debug output)
* `replay.min.js`(the minified bundle without debug output)
* `replay.debug.min.js` (the minified bundle with debug output)

For now, we'll only generate ES6 bundles but we can add ES5 versions if requested later on.