ref(nextjs): Use loader for rather than webpack plugin for injecting release (#6404)
This replicates the release-injection functionality provided by the webpack plugin using the prefix loader instead, in hopes of fixing https://github.com/getsentry/sentry-javascript/issues/5734, wherein the release value is somehow getting stuck on an old value. (My theory is that the file injected by the webpack plugin to inject the release is getting cached. This therefore switches to injecting the code setting the release using a loader, so that it's in the same file as, and just before, the `Sentry.init()` call.)