ref(replay): Streamline replay EventBuffer implementation (#6416)
This ensures public/private annotation is correct for the event buffer implementation. It also refactors a few (internal) things in minor ways, for clarity.

For example, we used a getter which as a side effect incremented itself. It is cleaner to do this explicitly, as getters should generally be side-effect free.