docs(replay): Add CDN bundle instructions to README (#6431)
Add a subsection about CDN bundle installation and usage. 

The only thing that's different in this doc compared to a normal integration CDN bundle doc on [Sentry docs](https://docs.sentry.io/platforms/javascript/install/cdn/#available-bundles) is that I can't specify the `integrity` attribute in the script tag because we can't automatically get the correct checksum in the Readme. I opted to leave it out and add a sentence that links people to our CDN bundle checksum table.