chore(replay): Remove Replay exclusion from Craft config (#6432)
Remove the exclusions of the @sentry/replay package from our craft config for the `github` and `npm` target. As specified in our checklist, we do not yet add a `registry` target entry for Replay as we'll have to add an entry manually to the release registry after this initial release.