fix(replay): Add CDN bundle path to release artifacts (#6452)
Craft uses the release artifacts zip generated by our "Upload Artifacts" job to search for files to upload. Our bundle path (`packages/replay/build/bundles`) wasn't added to this job's config so it wasn't included in the release.

This patch fixes that and it also adds an entry to our release checklist. It furthermore bumps the minimum version for the CDN bundles in the Replay README. Docs should automatically pick up the newest version, so we don't need to bump anything in https://github.com/getsentry/sentry-docs/pull/5899 

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>