ref(core): Move `getSdkMetadataForEnvelopeHeader` to utils (#6527)
We want to use `getSdkMetadataForEnvelopeHeader` in Replay to not have duplicated minimal SDKInfo extraction logic. I checked that envelope helper functions are usually publicly exposed from the utils package only (which makes sense to me). So, this PR moves the function from core to utils and calls it in the replay envelope creation function.

Note that I changed the input type to accept both, an `Event` as well as `SdkMetadata`, as both have the `sdk: SdkInfo` property. 

Since `getSdkMetadataForEnvelopeHeader` wasn't exported before, I think we don't have to worry about a breaking change here.