fix(replay): Send `dsn` in envelope header if tunneling is active (#6568)
As outlined in #6539, we were previosuly not sending the `dsn` key in the replay event [envelope header](https://develop.sentry.dev/sdk/envelopes/#envelope-headers), thereby not including the necessary information to forward the event to Sentry. 

This patch fixes that by extracting the `createEventEnvelopeHeaders` function to utils so that we can use it in Replay (analogously to #6514), as well as using this function to create all headers. 

We tested the change with NextJS' `tunnelRoute` as well as with the conventional `tunnel` option and everything seems to work now.