ref(types): Add `undefined` as possible event type (#6584)
Setting an `Event`'s property `type: undefined`, like we do in the SDK for `ErrorEvent`s causes a TS error, whenever the recommended `exactOptionalPropertyTypes` TS option is enabled. This option was introduced in TS 4.4. Since we're on TS 3.8, we couldn't catch this error. 

This patch changes to `EventType` type declaration to also accept `undefined` as a value, thereby resolving the TS error.