feat(replay): Streamline replay options
This streamlines options to be passed to replay a bit:

First, we deprecate passing arbitrary rrweb config through. Instead, you can provide `recordingOptions` directly, which will be passed through. For now, the old way still works, but we will remove this eventually.

Secondly, we explicitly defined the config options we _do_ support. This should cover most things that users are actually using, thus most users shouldn't need to actually do anything.

Finally, this also separates the options we pass to the integration from the options we store on the replay container. There is no need to keep all the rrweb-specific config there, we only need this at init-time, and can then discard it.