fix(replay): Generate uuids for replay events (#6650)
Instead of using the `replay_id`, we let sentry generate uuids for the `event_id` instead. Otherwise, multiple events will have the same `event_id`, which is not ideal. This still works because right now we do not actually use the `event_id` there, but for consistency it is better to have this be unique.