ci: Use nx cache from master by default
Currently, we do not store any cache for the master branch.
When we run a PR, it will try to restore the cache from itself, then from the current branch, then just the newest cache it finds at all.

This means that on a fresh branch, it will restore the cache from the last action that ran - which could be any branch, which means the cache may incorporate changed we don't have, leading to us having to rebuild the cache.

Instead, this makes sure we restore from the master branch first, if possible. This makes it necessary to ensure we actually also _store_ the cache when running on master - we only pass a restore key in that case that will never match anything.