fix(ci): Reduce flakiness of Replay integration tests (#6823)
Replace a bunch of `page.waitForTimeout` calls with `page.waitForRequest` waits which should make Replay integration tests more stable. 
Added a `waitForReplayRequest`, which ensures we're actually waiting for a replay request. Optionally, this helper takes a `segmentId` to wait for a replay request of a specific replay segment. This might be useful for more involved testing scenarios.