fix(replay): Only call `scope.getLastBreadcrumb` if available (#6969)
`Scope.getLastBreadcrumb` was introduced quite recently for replay and it generally works fine. Just in the few edge cases, where the base SDK version can't be changed or multiple hubs cause conflicts, Replay will crash when the method is not available on the passed scope. This patch lets `handleScope` early-return `null` in case the `getLastBreadcrumb` method does not exist. This will lead to missing breadcrumbs in the replay but at least it will avoid a crash.