feat(angular): Add Ivy-compatible Angular SDK package (#7264)
Add a new SDK package to our monorepo: `@sentry/angular-ivy`. 

While this is technically a new SDK, its content and functionality is identical to `@sentry/angular`. Only the build configuration differs:

* The Ivy SDK is built with Angular 12, allowing for a compatibility of NG12-15
* The Ivy SDK is built with `compilationMode: partial`, enabeling a build format that is compatible with Angular's Ivy rendering engine. 
  * This means that `ngcc` no longer needs to step in at initial build time to up-compile the SDK. Which is good because `ngcc` will be removed in Angular 16 (https://github.com/angular/angular-cli/pull/24720)
* The Ivy SDK's build output follows the Angular Package Format (APF) v12 standard ([spec](https://docs.google.com/document/d/1CZC2rcpxffTDfRDs6p1cfbmKNLA6x5O-NtkJglDaBVs/preview)) which is very similar to APF 10 which we used before (see #4641 for more details)

Because functionality-wise there's no difference to `@sentry/angular`, I opted to symlink the source files instead of duplicating them. The only exception is `sdk.ts` which needed some adaption for the new package, like setting the SDK name and adjusting the min version check. For the same reason, this new package currently doesn't contain tests. We'll need to reconsider this approach (symlinking and testing) if we ever need to make package-specific adjustments.