feat(replay): Reduce time limit before pausing a recording
Reduce the time limit before pausing a recording from `MAX_SESSION_LIFE` (1 hour) to `SESSION_IDLE_DURATION` (5 minutes). This reduces the amount of empty replays that are caused by a DOM mutation on a timer. An example of this is on sentry.io where there is a time component that updates every x seconds.

As a reminder, there are three events that will break the idle timeout:

* mouse click
* user input in a form field
* a navigation

Closes https://github.com/getsentry/sentry-javascript/issues/7352