feat(sveltekit): Add SvelteKit routing instrumentation (#7565)
Add routing instrumentation to the client SvelteKit SDK. Pageload navigations are created on function call as always and updated with a proper name once the `page` store emits the route id. Navigation transactions are created by subscribing to the `navigating` store.  No need for user configuration, as the instrumentation will be added automatically on SDK intialization.