fix(nextjs): Add tracing extension methods in `wrapServerComponentWithSentry` (#7567)
Our E2E tests were [failing](https://github.com/getsentry/sentry-javascript/actions/runs/4488942749/jobs/7894376591#step:6:6185) because apparently we don't initialize a client while building a NextJS app. This caused the tracing extension methods to not be added, resulting in `startTransaction` returning `undefined`. Not sure if this is the best way to fix this but it seems to work (h/t @timfish for the fix).