fix(sveltekit): Termporarily disable serverside load tracing (#7587)
We discovered that our serverside `handleLoadWithSentry` wrapper caused two distinct transactions being created on page loads - one for the page load request and one for the `load` function (if one exists for that page). This shouldn't happen. We need to investigate how to fix this. 

As a short-term solution, we disable tracing on the serverside load wrapper to avoid quota usage increase for our users.