fix(sveltekit): Handle same origin and destination navigations correctly (#7584)
Previously our Kit routing instrumentation used the parameterized route id to determine if the origin and destination of the navigation were identical (in which case we don't want to create a navigation transaction). 
This caused navigations from e.g. `/users/123` to `users/456` to not create a transaction because their route id was identical. This PR fixes this bug by using the raw URL which we also get from the `navigating` store emissions. 

Furthermore, this PR fixes the transaction source which previously was always set to `route`, even if no route id was available.