fix(sveltekit): Fix `<meta>` tag injection (#7612)
Apparently, the `%sveltekit.head%` placeholder isn't present in the html that's passed to `transformPageChunk`. Hence our `<meta>` tag injection logic introduced in #7574  didn't find the anchor point where to inject the tags. This PR fixes the injection, by simply appending it to an open `<head>` tag. It's not pretty but it should do the trick 🤞 (we can revisit if this breaks folks).