fix(sveltekit): Handle nested server calls in `sentryHandle` (#7598)
Previously, we always created a new domain when entering the `sentryHandle` function. This caused "nested" SvelteKit Server calls to create new transaction whenever a new domain was created in the `sentryHandle` function because the active span is bound to the domain. This fix introduces a check if there is an active transaction, in which case we just execute the handler and don't create a new domain, as we are already in this active domain. 

As a result, nested SK server calls become a span of the parent server call instead of a new transaction.