ref(sveltekit): Split up universal and server load wrappers (#7652)
 Split up the previously generic `handleLoadWithSentry` wrapper function into two functions on the server side:

* `handleLoadWithSentry` to wrap universal load on server and client
* `handleServerLoadWithSentry` to wrap server load on the server (duh)

It further adjusts types on the client as well as on the server side to return a generic. The very liberal typing of the functions' generic is on purpose because any type that's more narrowed down (e.g. `T extends Load`) would create a type error with the users' actual load types. These types (e.g. `PageData` or `ServerLayoutData`) are generated at build time per route/page. Hence, we cannot access them.