ref(sveltekit): Inject init plugin only if sentry config files exist (#7663)
To make our Sentry SDK setup easier, we'll instruct users to init the SDK in the hooks files rather than in the dedicated sentry.client|server.config.js files (#7661). To support this change, this PR makes the injection of our `injectSentryInitPlugin` into the users' build config depend on the existence of said sentry config files. If they don't exist, there's no need to add the plugin or adjust the server config.