fix(integration): Ensure HttpClient integration works with axios
This streamlines the integration to use the general fetch/xhr instrumentation, instead of adding it's own.

This also adds `request_headers` to the xhr info (which we'll also need for replay). 

It also ensures that we always correctly get the method/url from fetch options, even when they are other valid objects.