fix(node): Redact URL authority only in breadcrumbs and spans (#7740)
In #7667 we missed that our `urlToOptions` helper function is actually called to normalize request options that are then passed to the actual http client. Meaning, we shouldn't have redacted the authority in this function but at a later time when we extract the sanitized version (`extractUrl`). This PR changes the redaction location accordingly and hence fixes requests with authority not being sent properly. 

Co-authored-by: Luca Forstner <luca.forstner@sentry.io>