feat(sveltekit): Read adapter output directory from `svelte.config.js` (#7863)
Load and read the `svelte.config.js` file. This is necessary to automatically find the output directory that users can specify when setting up the Node adapter. 
This is a "little" hacky though, because we can't just access the output dir variable. Instead, we actually invoke the adapter (which we can access) and pass a minimal, mostly no-op adapter builder, which will report back the output directory.